[Feat]: 사진찍기 정보 로직 LocalDateUtil 함수로 분리
Related to: #33
