+++ module-domain/src/main/java/com/fourcut/diary/user/service/notification/NotificationTimeService.java
@@ -3,16 +3,19 @@ package com.fourcut.diary.user.service.notification;
 import com.fourcut.diary.user.domain.User;
 import com.fourcut.diary.user.domain.notification.NotificationTime;
 import com.fourcut.diary.user.service.dto.TakePhotoInfoDto;
-import com.fourcut.diary.user.util.NotificationTimeUtil;
+import com.fourcut.diary.util.LocalDateTimeUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class NotificationTimeService {
+
+    private static final int EXPIRATION_MINUTES = 20;
 
     private final NotificationTimeRetriever notificationTimeRetriever;
 
@@ -20,6 +23,23 @@ public class NotificationTimeService {
     public TakePhotoInfoDto getTakePhotoInfoByUser(User user) {
         NotificationTime notificationTime = notificationTimeRetriever.findNotificationTimeByUser(user);
         LocalDateTime now = LocalDateTime.now();
-        return NotificationTimeUtil.getTakePhotoInfo(notificationTime, now);
+        return getTakePhotoInfo(notificationTime, now);
+    }
+
+    private TakePhotoInfoDto getTakePhotoInfo(NotificationTime notificationTime, LocalDateTime now) {
+        List<LocalDateTime> timeSlots = List.of(
+                notificationTime.getFirstTimeSlot(),
+                notificationTime.getSecondTimeSlot(),
+                notificationTime.getThirdTimeSlot(),
+                notificationTime.getFourthTimeSlot()
+        );
+
+        for (int i = 0; i < timeSlots.size(); i++) {
+            LocalDateTime slot = timeSlots.get(i);
+            if (LocalDateTimeUtil.getIsPossiblePhotoCapture(slot, now)) {
+                return new TakePhotoInfoDto(i + 1, slot.plusMinutes(EXPIRATION_MINUTES).toLocalTime());
+            }
+        }
+        return new TakePhotoInfoDto(-1, null);
     }
 }
